Feature/fix coredata in release-2.11.0

+++ GroceryList/ViewControllers/Recipe screens/MainRecipeViewController/MainRecipeDataSource.swift
@@ -131,8 +131,10 @@ class MainRecipeDataSource: MainRecipeDataSourceProtocol {
     }
     
     private func removeDefaultsCollection() {
-        EatingTime.defaults.forEach {
-            CoreDataManager.shared.deleteCollection(by: $0.rawValue)
+        if !UserDefaultsManager.shared.isFillingDefaultCollection {
+            EatingTime.defaults.forEach {
+                CoreDataManager.shared.deleteCollection(by: $0.rawValue)
+            }
         }
     }
